Work around qualified identifier issue in Isabelle generation
Using the qualified name `InstructionSemantics.decode_to_instruction` of
a field in the Lem source will generate the same qualified name for
Isabelle, but Isabelle expects the name of the type here instead of (or
in addition to) the name of the source theory where the record type is
defined.  Unqualified names work here, though, so use those.